# Performance: Adds AuthorizationHelper improvements